Refatora ProductsController e as funções de model para uma estrutura baseada em classes. Cria ProductsService com os métodos responsáveis pela comunicação com o banco de dados.
